# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie App API",
        "description": "Tài liệu API cho hệ thống đặt vé xem phim",
        "contact": {
            "email": "youremail@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/suatchieu/phim/{ma_phim}": {
            "get": {
                "tags": [
                    "Suất Chiếu"
                ],
                "summary": "Lấy danh sách suất chiếu theo mã phim",
                "description": "API này dùng để lấy tất cả suất chiếu cho một phim.",
                "operationId": "2a0249d35ac0b9a512c4965c180546c6",
                "parameters": [
                    {
                        "name": "ma_phim",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "PH001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách suất chiếu của phim",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ma_rap": {
                                            "type": "string",
                                            "example": "RP001"
                                        },
                                        "ten_rap": {
                                            "type": "string",
                                            "example": "Rạp Phim ABC"
                                        },
                                        "dia_chi": {
                                            "type": "string",
                                            "example": "123 Đường ABC, Quận 1, TP.HCM"
                                        },
                                        "suat_chieu": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ma_suat_chieu": {
                                                        "type": "string",
                                                        "example": "SC001"
                                                    },
                                                    "ma_phong": {
                                                        "type": "string",
                                                        "example": "P001"
                                                    },
                                                    "phong": {
                                                        "type": "string",
                                                        "example": "Phòng 1"
                                                    },
                                                    "thoi_gian_bd": {
                                                        "type": "string",
                                                        "example": "10:00"
                                                    },
                                                    "ngay_chieu": {
                                                        "type": "string",
                                                        "example": "12-04-2025"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không có suất chiếu cho phim này",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Không có suất chiếu cho phim này"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}